backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-core
langchain-text-splitters
python-multipart
google-genai
